fw/arcar/.scripts/requirements.txt
colorama>=0.3.9
pyserial>=3.4
